src/YOLOApp/requirements.txt
aiohttp
aiortc
asyncio
aiohttp_rest_api
tensorflow==2.2
Pillow